refactor(magicalitem): move AddMagicalItem* classes to userlist.presentation package
