+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/userlist/presentation/AddMagicalItemToListState.kt
@@ -1,10 +1,10 @@
 package com.cyrillrx.rpg.userlist.presentation
 
-import com.cyrillrx.rpg.spell.domain.Spell
+import com.cyrillrx.rpg.magicalitem.domain.MagicalItem
 import com.cyrillrx.rpg.userlist.domain.UserList
 import org.jetbrains.compose.resources.StringResource
 
-data class AddToListState(
+data class AddMagicalItemToListState(
     val body: Body = Body.Loading,
 ) {
     data class SelectableUserList(val list: UserList, val alreadyAdded: Boolean, val isSelected: Boolean = alreadyAdded)
@@ -13,7 +13,7 @@ data class AddToListState(
         data object Loading : Body
         data class Error(val errorMessage: StringResource) : Body
         data class WithData(
-            val spell: Spell,
+            val magicalItem: MagicalItem,
             val selectableLists: List<SelectableUserList>,
         ) : Body
     }
